reorganize all of the project files
