+++ apps/server/package.json
@@ -14,8 +14,6 @@
   },
   "dependencies": {
     "commander": "^13.1.0",
-    "elysia": "^1.2.25",
-    "yaml": "^2.7.1",
     "zod": "^3.24.2",
     "inksync-sdk": "workspace:*"
   }
