--- apps/server/src/main.ts
@@ -1,3 +1,0 @@
-import { runCli } from "./cli";
-
-runCli();
