+++ packages/inksync-sdk/server/auth.ts
@@ -1,4 +1,3 @@
-
 const validTokens = ["token1", "token2"];
 
 export function tokenIsValid(token: string): boolean {
